python/requirements.txt
blinker==1.4
Click==7.0
dnspython==1.16.0
Flask==1.1.1
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.1
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
mysql-connector-python==8.0.18
protobuf==3.10.0
six==1.13.0
SQLAlchemy==1.3.10
uWSGI==2.0.18
Werkzeug==0.16.0
